requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
werkzeug==2.3.7
gunicorn==20.1.0